Update game.cpp
finished the initial trade_with_player function. added comment that is significant immediately before trade_with_player...
it is the server should probably check if the requested player even has the cards before sending them the trade request. then the server can auto deny the trade. Or, dont and let the code check after they approve the trade. then if the trade is invalid but approved, the server can respond with a message saying the trade is invalid. 

Either implementation works, but the later option may be better so that a player cant send a bunch of trades in very quickly to see what cards each player has, since one could assume a response within a very short amount of time is the server rejecting the trade