Update server_Catan.cpp
added comments to todo list. need to start working on the game itself and not on the communications as much. once the game play is mostly done, revisit the communcation. game play meaning how turns are handled and everything, all of which is built off of game.cpp and stuff. this file should only handle the communication and basic tasks, and then return some value back to the main game loop of the sevrer for further action (ex: when start turn is received, the framehandler could either return a code saying to start the turn, OR it could call start turn and then return the dice roll. may be better to not have it do that and only have it return actions to do with some other info so that framehandler is available for other incoming packets. have flags to pass other data and states back and forth.