think I find a solution, see description. (tile_client.cpp)
adding this-> didnt work to the front of clientcornersz name didnt work,
so I think the issue is that clientBaton.cpp initializes the tilecorners
class, but doesnt present it to catan baton. I should change
client_baton.cpp into something similar to game.cpp for the client. it
should contain everything in a class and then allow the user to start a
new game by defining a new class!