added the server project for stand alone server
the server project file is now located under Servergame, but the code is
still under project2 folder. The servergame project has the stand alone
server code to run, and is now stand alone. there is still more
functionality to be added to the server code, and the tcp stuff needs to
be worked out, but it is successfully opening the server ports and what
not.